log-capture-service/requirements.txt
watchdog==4.0.1
pyyaml==6.0.2
